refactor: refatora o ZipCodeUtils.
